requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
openai==1.3.8
